Add Docker setup for production preview
This commit introduces Docker configuration to the project strictly for production preview purposes. The update includes creation of Dockerfile, docker-compose.yml, .dockerignore files and modifications to .gitignore and next.config.mjs files. The Dockerfile contains build instructions and setup for a Next.js application and the docker-compose provides container orchestration. Important to note, local development should not be done using Docker. It only serves the purpose of previewing the production environment.