Add theme toggle feature and update dependencies
This commit introduces a new feature that allows users to toggle between light, dark, and system themes. This is implemented in the `ThemeToggle` component. Additionally, the `GlobalProviders` component is introduced as a context provider for the application.

The `react-icons` package (version 5.0.1) was added to the `package.json` and `pnpm-lock.yaml` files to support the new theme switching functionality. The corresponding dependency is also updated in the lock file for consistency.

Furthermore, TypeScript types for layout and pages were abstracted into the `common-types.ts` file for reusability and better code structure.

Lastly, these changes have been integrated into the `layout.tsx` file. This involved including the `GlobalProviders` and `ThemeToggle` components and updating the props type for `RootLayout` to `LayoutType`.